+++ rust/main/agents/scraper/src/agent.rs
@@ -743,11 +743,14 @@ mod test {
                 metrics_conf: PrometheusMiddlewareConf {
                     contracts: HashMap::new(),
                     chain: None,
+                    rpc_role: Default::default(),
                 },
                 index: IndexSettings {
                     from: 0,
                     chunk_size: 1,
                     mode: IndexMode::Block,
+                    idle_sleep_duration: Duration::from_secs(5),
+                    configured_interval: None,
                 },
                 confirmations: Default::default(),
                 chain_id: Default::default(),