+++ typescript/utils/src/index.ts
@@ -243,7 +243,9 @@ export {
   isSdkSupportedProtocol,
   TokenCaip19Id,
   ValidatorMetadata,
+  ValidatorMetadataRpcEntry,
   WithAddress,
+  validatorMetadataRpcUrlHash,
 } from './types.js';
 export { isHttpsUrl, isRelativeUrl, isUrl } from './url.js';
 export { assert } from './validation.js';